Disallow registering KeyManagers for primitives other than Aead, DeterministicAead, StreamingAead, HybridEncrypt, HybridDecrypt, Mac, PublicKeySign, PublicKeyVerify, Prf.
We plan to refactor some of the internals, and for KeysetDeriver things will likely break. For JWT it doesn't really make sense if the user adds their own primitives. I also think it's extremely unlikely that someone uses a custom type, so I'd suggest to just disable registering of these types.

If I am guessing wrong and users use this with different types, I would ask them to file an issue on https://github.com/tink-crypto/tink-java so that we can find the best solution.

PiperOrigin-RevId: 569424461