Move newKeyAllowed handling into the KeyManagerRegistry.
I want to remove KeyTypeManagers as they should -- with the new key types as java classes -- not be needed anymore. However, for this I need to make a new mechanism for the FIPS registration. This will work as follows: the KeyManagerRegistry needs to provide an internal method "registerFipsKeyManager".

However, for this, in turn, we need to make sure that if people call register in parallel in Registry and in KeyManagerRegistry we have no concurrency issues.

For this, we need then to ensure that everything can be done in the KeyManagerRegistry, and the registry just forwards to it. This is the first step in this direction.

PiperOrigin-RevId: 574456716