Remove publicKeyManagerClassOrNull and the corresponding check from the KeyManager registry.
This check used to check that if you do two calls:
  registerAsymmetricKeyManagers(privateKeyManager, publicKeyManagerA);
  registerAsymmetricKeyManagers(privateKeyManager, publicKeyManagerB);
then it fails.

However, this isn't really doing much: if one does so, it anyhow already fails since probably publicKeyManagerA and publicKeyManagerB have the same typeURL -- and otherwise it's not even important. It seems this just makes everything more complicated.

PiperOrigin-RevId: 580446347