Remove the Registry methods which have a need the "TypeParameterUnusedInFormals" warning suppression.
These are a misuse of java generics (see https://errorprone.info/bugpattern/TypeParameterUnusedInFormals). Since they have been deprecated for 5 years, we will remove them now.

Since these methods are in the registry it seems safe to assume that nobody uses these.

The existence of these functions implies that previously the following code would compile and run as indicated:

static void fun() throws Exception {
  KeyManager<Mac> myKeyManager = createSomeMacKeyManager();
  String typeUrl = myKeyManager.getKeyType();
  Registry.registerKeyManager(myKeyManager);
  KeyManager<Aead> aeadKeyManager = registry.getKeyManager(typeUrl);
  // At this point, aeadKeyManager == myKeyManager holds.
  // This is even though aeadKeyManager isn't even of the correct type.
  // However, the generic type information isn't present in Java at runtime,
  // and we successfully tricked the compiler to ignored this issue.
}

Users who use this function need to add the generic argument to it -- in the above case, add "Mac.class" after typeUrl in getKeyManager.

The reason I want to remove it is because it becomes increasingly tricky to maintain previous behavior. For example, we soon should have the ability to register multiple key manager with the same type url -- and these methods need to somehow guess the correct key manager to return.

PiperOrigin-RevId: 581153043