Factor out a method "GetMatchedStream".
This simplifies the logic, but on the first call, this slightly slows down the reading (because if we want to read, we first call size to provoke the matching, which reads at least some part of the ciphertext. afterwards we read ciphertext from the start again).

This also fixes the bug that calling size() would fail if PRead isn't called before this.

PiperOrigin-RevId: 584556418