Replace instances of Registry.registerPrimitiveWrapper with MutablePrimitiveRegistry.globalInstance().registerPrimitiveWrapper.
Registry.registerPrimitiveWrapper will most likely be removed. The API only works if used correctly together with what is in KeysetHandle. Furthermore, we are currently changing this, so if users use this API their code will inevitably fail.

I don't expect anyone to use this, otherwise they should contact us and we will find a solution together. (I will add this desclaimer in the upcoming change which removes this as well).

PiperOrigin-RevId: 601441934