Use different AEADs for different key names in FakeHcVault
When creating FakeHcVault the user must provide a mount path and a list of valid key names, for each of which the fake implementation creates a new AEAD with a different key. This makes its behavior more realistic:
(1) different key names will correspond to different keys (i.e., AEADs);
(2) the fake implementation fails if the mount path is invalid and/or an unknown key is specified.

And add unit test to getOperationEndpoint, and fix implementation.

PiperOrigin-RevId: 604719591