Avoid registry functions Registry.newKeyData and Registry.getPrimitive in the KmsEnvelopeAead.
These two functions should not be in the public API and I will deprecate them. In this particular case we need to go into internals somewhat to do the same thing. This can be a problem (since it means users cannot do the same thing anymore). However, I think here this is on purpose: the class actually have problems in behavior (it just parses a dek with some potentially wrong proto), and so there will be no official API to do what it does.

If users face similar problems we will have to find solutions with them individually.

PiperOrigin-RevId: 609259134