Check for FIPS compatibility at beginning of JWT HMAC key manager registration.
This makes sure that nothing gets registered if Tink is used in FIPS mode but the FIPS module is not available.

Also, the documentation of registerKeyManagerWithFipsCompatibility says that if it fails, then "the KeyManagerRegistry is in an unspecified state and should be discarded." But since we are calling it on a global instance that can't be discarded, we need to make sure that the call doesn't fail. And if it does fail, we need to throw a runtime exception as this shouldn't happen.

PiperOrigin-RevId: 611373559