Remove AndroidKeystoreKmsClient.setKeyStore from public API.
This function is marked for testing only, and was used to inject a fake KeyStore instance. But it doesn't work as expected, because AndroidKeystoreKmsClient.hasKey may overwrite this with a real KeyStore instance. And the static functions always use the real KeyStore. I think this is confusing and it's better to completely remove this method. If you need to test your code with a fake KeyStore instance, it is preferable to inject fake security provider using Security.addProvider, see FakeAndroidKeystoreProvider.java as an example for such a fake security provider.

PiperOrigin-RevId: 612810146